wip; builder tests green, instr moved to models/instrumentation
