+++ lib/experiments/index.js
@@ -3,6 +3,7 @@ var symbol = require('../utilities/symbol');
 
 var experiments = {
   BUILD_INSTRUMENTATION: symbol('build-instrumentation'),
+  INSTRUMENTATION: symbol('instrumentation'),
   ADDON_TREE_CACHING: symbol('addon-tree-caching')
 };
