+++ lib/models/builder.js
@@ -11,75 +11,9 @@ var SilentError = require('silent-error');
 var chalk       = require('chalk');
 var attemptNeverIndex = require('../utilities/attempt-never-index');
 var findBuildFile = require('../utilities/find-build-file');
-var experiments = require('../experiments');
 var _resetTreeCache = require('./addon')._resetTreeCache;
 
 var Sync = require('tree-sync');
-
-
-function _computeVizInfo(count, result, resultAnnotation) {
-  var graph = require('heimdalljs-graph');
-  var buildTree = graph.loadFromNode(result.graph.__heimdall__);
-  var buildSteps = 0;
-  var totalTime = 0;
-
-  var node;
-  var statName;
-  var statValue;
-  var nodeItr;
-  var statsItr;
-  var nextNode;
-  var nextStat;
-
-  for (nodeItr = buildTree.dfsIterator();;) {
-    nextNode = nodeItr.next();
-    if (nextNode.done) { break; }
-
-    node = nextNode.value;
-    if (node.label.broccoliNode && !node.label.broccoliCachedNode) {
-      ++buildSteps;
-    }
-
-    for (statsItr = node.statsIterator();;) {
-      nextStat = statsItr.next();
-      if (nextStat.done) { break; }
-
-      statName = nextStat.value[0];
-      statValue = nextStat.value[1];
-
-      if (statName === 'time.self') {
-        totalTime += statValue;
-      }
-    }
-  }
-
-  var vizJson = {
-    /* we should calculate the summary more generically */
-    summary: {
-      build: {
-        type: resultAnnotation.type,
-        count: count,
-        outputChangedFiles: result.outputChanges,
-      },
-      output: result.directory,
-      totalTime: totalTime,
-      buildSteps: buildSteps,
-    },
-    buildTree: buildTree,
-  };
-
-  if (resultAnnotation.type === 'rebuild') {
-    vizJson.summary.build.primaryFile = resultAnnotation.primaryFile;
-    vizJson.summary.build.changedFileCount = resultAnnotation.changedFiles.length;
-    vizJson.summary.build.changedFiles = resultAnnotation.changedFiles.slice(0, 10);
-  }
-
-  return vizJson;
-}
-
-function writeVizInfo(count, vizJson) {
-  fs.writeJsonSync('broccoli-viz.' + (count) + '.json', vizJson);
-}
 
 /**
  * Wrapper for the Broccoli [Builder](https://github.com/broccolijs/broccoli/blob/master/lib/builder.js) class.
@@ -94,7 +28,6 @@ module.exports = CoreObject.extend({
 
   init: function() {
     this._super.apply(this, arguments);
-    this._buildCount = 0;
     this.setupBroccoliBuilder();
     this.trapSignals();
     this._instantiationStack = (new Error).stack.replace(/[^\n]*\n/, '');
@@ -267,25 +200,14 @@ module.exports = CoreObject.extend({
     });
   },
 
-  _reportVizInfo: function (vizInfo) {
-    if (!experiments.BUILD_INSTRUMENTATION) { return; }
-
-    if (this.project && this.project.addons.length) {
-      this.project.addons.forEach(function (addon) {
-        var hook = addon[experiments.BUILD_INSTRUMENTATION];
-        if (typeof hook === 'function') {
-          hook.call(addon, vizInfo);
-        }
-      });
-    }
-  },
-
   /**
    * @private
    * @method build
    * @return {Promise}
    */
   build: function(willReadStringDir, resultAnnotation) {
+    this.project.instrumentation.start('build');
+
     var self = this;
     attemptNeverIndex('tmp');
 
@@ -296,26 +218,7 @@ module.exports = CoreObject.extend({
       .then(this.processAddonBuildSteps.bind(this, 'postBuild'))
       .then(this.processBuildResult.bind(this))
       .then(function(result) {
-        var vizJson;
-
-        if (self.project.instrumentation.isEnabled()) {
-          vizJson = self._computeVizInfo(self._buildCount, result, resultAnnotation);
-          self._reportVizInfo(vizJson);
-        }
-
-        if (self.project.instrumentation.isVizEnabled()) {
-          vizJson.nodes = result.graph.__heimdall__.toJSONSubgraph();
-          writeVizInfo(self._buildCount,{
-            summary: vizJson.summary,
-            // We want to change this to buildTree, to be consistent with the
-            // hook, but first we must update broccoli-viz
-            // see https://github.com/ember-cli/broccoli-viz/issues/35
-            nodes: result.graph.__heimdall__.toJSONSubgraph()
-          });
-        }
-
-        self._buildCount++;
-
+        self.project.instrumentation.stopAndReport('build', result, resultAnnotation);
         return result;
       })
       .then(this.processAddonBuildSteps.bind(this, 'outputReady'))
@@ -405,8 +308,4 @@ module.exports = CoreObject.extend({
       process.exit(1);
     }
   },
-
-  _computeVizInfo: _computeVizInfo,
 });
-
-module.exports._computeVizInfo = _computeVizInfo;
